Replace UniswapV3Oracle tests with UniswapV3PriceAccumulator tests
